Added Docker Files and Comments
