+++ .env
@@ -1,0 +1,5 @@
+# .env
+COMPOSE_PROJECT_NAME=ecommerce
+POSTGRES_DB=ECommerceDB
+POSTGRES_USER=postgres
+POSTGRES_PASSWORD=QWERTY
