+++ .gitignore
@@ -9,3 +9,10 @@
 /frontend/.idea/
 /frontend/.react-router/
 /frontend/node_modules/
+
+
+# Docker
+.env.local
+.env.prod
+*.env
+docker/*/*.env
