+++ backend/backend.http
@@ -1,4 +1,4 @@
-@backend_HostAddress = http://localhost:5026
+@backend_HostAddress = http://localhost:8080
 
 GET {{backend_HostAddress}}/weatherforecast/
 Accept: application/json
