+++ backend/src/Controllers/CategoryController.cs
@@ -1,11 +1,14 @@
 using Microsoft.AspNetCore.Mvc;
 using Models;
 using Services;
+using Microsoft.AspNetCore.Authorization;
+using System.Security.Claims;
 
 namespace Controllers;
 
 [ApiController]
 [Route("api/[controller]")]
+[Authorize]
 public class CategoryController : ControllerBase {
     private readonly CategoryService _service;
     public CategoryController(CategoryService service) {
@@ -41,12 +44,24 @@ public class CategoryController : ControllerBase {
 
     [HttpPost("Add")]
     public async Task<IActionResult> CreateCategory(Category category) {
+        var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        // verify user role
+        if (!User.IsInRole("ADMIN"))
+        {
+            return Forbid();
+        }
         var id = await _service.CreateAsync(category);
         return CreatedAtAction(nameof(GetCategories), new { id }, category);
     }
     
     [HttpPut("Update-by-ID/{id}")]
     public async Task<IActionResult> UpdateCategory(int id, Category category) {
+        var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        // verify user role
+        if (!User.IsInRole("ADMIN"))
+        {
+            return Forbid();
+        }
         if (id != category.Id)
             return BadRequest("Category ID mismatch.");
 
@@ -56,6 +71,12 @@ public class CategoryController : ControllerBase {
     
     [HttpDelete("Delete-by-ID/{id}")]
     public async Task<IActionResult> DeleteCategory(int id) {
+        var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        // verify user role
+        if (!User.IsInRole("ADMIN"))
+        {
+            return Forbid();
+        }
         await _service.DeleteAsync(id);
         return NoContent();
     }
