+++ backend/src/Controllers/UserController.cs
@@ -1,4 +1,6 @@
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Authorization;
+using System.Security.Claims;
 using Models;
 using Services;
 
@@ -6,6 +8,7 @@ namespace Controllers
 {
     [ApiController]
     [Route("api/[controller]")]
+    [Authorize]
     public class UserController : ControllerBase
     {
         private readonly UserService _service;
@@ -18,6 +21,12 @@ namespace Controllers
         [HttpGet("Get")]
         public async Task<IActionResult> GetAllUsers()
         {
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            // verify user role
+            if (!User.IsInRole("ADMIN"))
+            {
+                return Forbid();
+            }
             var users = await _service.GetAllUsersAsync();
             return Ok(users);
         }
@@ -25,6 +34,12 @@ namespace Controllers
         [HttpGet("Get/count")]
         public async Task<IActionResult> GetTotalUsersCount()
         {
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            // verify user role
+            if (!User.IsInRole("ADMIN"))
+            {
+                return Forbid();
+            }
             var count = await _service.GetTotalUsersAsync();
             return Ok(new { totalUsers = count });
         }
