+++ backend/src/Model/Auth.cs
