+++ backend/src/Model/Product.cs
