+++ backend/src/Repositories/ProductRepository.cs
@@ -23,7 +23,6 @@ public class ProductRepository {
                 Id = reader.GetInt32(0),
                 Name = reader.GetString(1),
                 Description = reader.IsDBNull(2) ? null : reader.GetString(2),
-                // Price = reader.GetDouble(3),
                 Price = reader.GetDecimal(reader.GetOrdinal("price")),
                 StockQuantity = reader.GetInt32(4),
                 CategoryId = reader.GetInt32(5),
