+++ package.json
@@ -1,0 +1,9 @@
+{
+  "scripts": {
+    "docker:dev": "docker-compose -f docker-compose.yml -f docker/development/docker-compose.dev.yml --env-file docker/development/.env.dev up",
+    "docker:prod": "docker-compose -f docker-compose.yml -f docker/production/docker-compose.prod.yml --env-file docker/production/.env.prod up -d",
+    "docker:down": "docker-compose down",
+    "docker:build": "docker-compose build",
+    "docker:logs": "docker-compose logs -f"
+  }
+}
